Cambios gise ver2:

Sacar progress Bar en Buscar cuando no hay resultados
Arreglo refresh en home que no estaba andando
Agrego toolbar para el detalle del articulo
Oculto rating bar para las publicaciones del destacado
Arrelo en las solicitudes cerradas, para que el label dig que estan c… …
Volver en tienda
Agrego soporte a varias fotos
Soporte vidriera
Cambios en el fragmen contact de la tienda y arreglo de ubicacion
Limpio lista cuando me deslogueo, intento de sacar ads de las cards
Rating de recomendaciones en tienda y bug del search
Fotos en conjuntos y validacion para conjuntos vacios

+++ mobile/app/src/main/java/com/mylook/mylook/entities/Store.java
@@ -17,8 +17,8 @@ public class Store {
     private String storeAddress;
     private String storeAddressNumber;
     private String storeFloor;
-    private long storeLatitude;
-    private long storeLongitude;
+    private double storeLatitude;
+    private double storeLongitude;
     private String storeDescription;
     private String instagramLink;
     private String twitterLink;
@@ -177,19 +177,19 @@ public class Store {
         this.twitterLink = twitterLink;
     }
 
-    public long getStoreLatitude() {
+    public double getStoreLatitude() {
         return storeLatitude;
     }
 
-    public void setStoreLatitude(long storeLatitude) {
+    public void setStoreLatitude(double storeLatitude) {
         this.storeLatitude = storeLatitude;
     }
 
-    public long getStoreLongitude() {
+    public double getStoreLongitude() {
         return storeLongitude;
     }
 
-    public void setStoreLongitude(long storeLongitude) {
+    public void setStoreLongitude(double storeLongitude) {
         this.storeLongitude = storeLongitude;
     }
